backend/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
pymongo==3.10.1
python-jose==3.2.0
passlib==1.7.4
pytz==2019.3
XlsxWriter==3.0.1